requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
mypy-extensions==0.4.3
numpy==1.19.5
pandas==1.2.1
pathspec==0.8.1
praw==7.1.0
prawcore==1.5.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.3
websocket-client==0.57.0
